fix bugs

+++ lab0/utils.h
@@ -36,7 +36,7 @@ int  init_GLSL(char * vert_source, char * frag_source);
 
 jmesh * readOffFile(const char * filename);
 void printMesh(const jmesh *);
-void drawCube2(int speed, float dist, int index);
+void drawCube2(float speed, float dist, int index);
 void drawTest();
 void myTimerFunc(int val);
